+++ ng-aruco-marker.js
@@ -1,7 +1,23 @@
 // AngularJS Directive for Aruco marker images
 // Usage:
 // <aruco-marker marker-id="{{markerId}}" size="500px"/>
-(function() {
+(function (root, factory) {
+  if (typeof module !== 'undefined' && module.exports) {
+    // CommonJS
+    if (typeof angular === 'undefined') {
+      factory(require('angular'));
+    } else {
+      factory(angular);
+    }
+    module.exports = 'ngDialog';
+  } else if (typeof define === 'function' && define.amd) {
+    // AMD
+    define(['angular'], factory);
+  } else {
+    // Global Variables
+    factory(root.angular);
+  }
+}(this, function (angular) {
   angular.module('arucoMarker', []).directive('arucoMarker', function() {
     return {
       restrict: 'E',
@@ -17,4 +33,4 @@
       }
     };
   });
-})();
+}));