+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-0.1.0
+2.0.0
+-----
+
+* Fix marker orientation - it was diagonally inverted!
+* Add CommonJS and AMD support to the Angular directive.
+
+1.0.0
 -----
 
 * Initial release